+++ src/KurrentDB.UI/Layout/MainLayout.razor
@@ -1,12 +1,12 @@
-﻿@using System.Security.Claims
+﻿@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
+@using System.Security.Claims
 @using KurrentDB.UI.Services
 @inherits LayoutComponentBase
 @inject Preferences Preferences
 
 <MudThemeProvider @ref="@_mudThemeProvider" Theme="@_theme" IsDarkMode="Preferences.DarkMode"/>
-<MudPopoverProvider/>
-<MudDialogProvider/>
-<MudSnackbarProvider/>
 <MudLayout>
 	<MudAppBar Elevation="1">
 		<MudIconButton Icon="@Icons.Material.Filled.Menu" Color="Color.Inherit" Edge="Edge.Start"
@@ -46,19 +46,6 @@
 
 		Preferences.SetTheme(await _mudThemeProvider.GetSystemPreference());
 		StateHasChanged();
-		// AuthService.UserChanged += HandleUserChange;
-		//
-		// if (AuthService.IsLoggedIn) return;
-
-		// var url = NavigationManager.ToAbsoluteUri(NavigationManager.Uri).ToString();
-		// if (url.Contains("/login")) {
-		// 	return;
-		// }
-
-		// var restoredFromState = await AuthService.GetStateFromTokenAsync();
-		// if (restoredFromState) {
-		// 	NavigationManager.Refresh();
-		// }
 	}
 
 	void HandleUserChange(ClaimsPrincipal principal) {
