+++ src/KurrentDB.UI/Layout/NavMenu.razor
@@ -1,4 +1,7 @@
-﻿<MudNavMenu>
+﻿@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
+<MudNavMenu>
 	<MudNavLink Href="/ui/cluster" Match="NavLinkMatch.All" Icon="@Icons.Material.Filled.Thermostat">Status</MudNavLink>
 	<MudNavLink Href="/ui/logs" Match="NavLinkMatch.All" Icon="@Icons.Material.Filled.ManageSearch">Logs</MudNavLink>
 	<MudNavLink Href="/ui/config" Match="NavLinkMatch.All" Icon="@Icons.Material.Filled.Settings">Config</MudNavLink>
