+++ src/KurrentDB.UI/RedirectToLogin.razor
@@ -1,4 +1,7 @@
-﻿@inject NavigationManager NavigationManager
+﻿@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
+@inject NavigationManager NavigationManager
 
 @code {
 	protected override void OnInitialized() {
