+++ src/KurrentDB.UI/Services/PersistentAuthenticationStateProvider.cs
@@ -1,17 +1,12 @@
+// Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements.
+// Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md).
+
 using System.Security.Claims;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Authorization;
 
 namespace KurrentDB.UI.Services;
 
-// This is a client-side AuthenticationStateProvider that determines the user's authentication state by
-// looking for data persisted in the page when it was rendered on the server. This authentication state will
-// be fixed for the lifetime of the WebAssembly application. So, if the user needs to log in or out, a full
-// page reload is required.
-//
-// This only provides a username and email for display purposes. It does not actually include any tokens
-// that authenticate to the server when making subsequent requests. That works separately using a
-// cookie that will be included on HttpClient requests to the server.
 class PersistentAuthenticationStateProvider : AuthenticationStateProvider {
 	static readonly Task<AuthenticationState> DefaultUnauthenticatedTask = Task.FromResult(new AuthenticationState(new(new ClaimsIdentity())));
 
@@ -22,10 +17,7 @@ class PersistentAuthenticationStateProvider : AuthenticationStateProvider {
 			return;
 		}
 
-		Claim[] claims = [
-			new(ClaimTypes.NameIdentifier, userInfo.UserId),
-			new(ClaimTypes.Name, userInfo.UserId),
-		];
+		Claim[] claims = [new(ClaimTypes.NameIdentifier, userInfo.UserId), new(ClaimTypes.Name, userInfo.UserId)];
 
 		_authenticationStateTask = Task.FromResult(new AuthenticationState(new(new ClaimsIdentity(claims, authenticationType: nameof(PersistentAuthenticationStateProvider)))));
 	}
