+++ src/KurrentDB/Components/App.razor
@@ -1,4 +1,7 @@
-﻿@using Microsoft.AspNetCore.Http
+﻿@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
+@using Microsoft.AspNetCore.Http
 @using KurrentDB.UI
 <!DOCTYPE html>
 <html lang="en">
