+++ src/KurrentDB/Components/Auth/Login.razor
@@ -1,58 +1,67 @@
+@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
 @page "/ui/login"
+@attribute [ExcludeFromInteractiveRouting]
 @using System.ComponentModel.DataAnnotations
 @using System.Security.Claims
-@using EventStore.Core.Authentication
 @using EventStore.Plugins.Authentication
 @using KurrentDB.Services
+@using Microsoft.AspNetCore.Authentication
 @using Microsoft.AspNetCore.Http
+@using MudBlazor.StaticInput
 
-<EditForm Model="Input" OnValidSubmit="LoginUser" FormName="LoginUser">
-	<DataAnnotationsValidator/>
-	<MudGrid>
-		<MudItem xs="6">
-			<MudCard>
-				<MudCardContent>
-					<MudTextField Label="Username" HelperText="Enter your username" AutoFocus="true"
-					              @bind-Value="Input.Username" For="@(() => Input.Username)"/>
-					<MudTextField Label="Password" HelperText="Enter your password" Class="mt-3"
-					              @bind-Value="Input.Password" For="@(() => Input.Password)"
-					              InputType="InputType.Password"/>
-				</MudCardContent>
-				<MudCardActions>
-					<MudButton ButtonType="ButtonType.Submit" Variant="Variant.Filled" Color="Color.Primary"
-					           Class="ml-auto">Log in
-					</MudButton>
-				</MudCardActions>
-			</MudCard>
-		</MudItem>
-		<MudFlexBreak/>
-		<MudItem xs="6">
-		</MudItem>
-	</MudGrid>
-</EditForm>
+<PageTitle>Log in</PageTitle>
+
+<MudGrid>
+	<MudItem md="6">
+		@* <StatusMessage Message="@errorMessage"/> *@
+		<EditForm Model="Input" method="post" OnValidSubmit="LoginUser" FormName="login">
+			<DataAnnotationsValidator/>
+
+			<MudText GutterBottom="true" Typo="Typo.body1">Use a local account to log in.</MudText>
+
+			<MudGrid>
+				<MudItem md="12">
+					<MudStaticTextField For="@(() => Input.Username)" @bind-Value="Input.Username"
+					                    Label="Username" Placeholder="admin"
+					                    UserAttributes="@(new() { { "autocomplete", "username" }, { "aria-required", "true" } })"/>
+				</MudItem>
+				<MudItem md="12">
+					<MudStaticTextField For="@(() => Input.Password)" @bind-Value="Input.Password"
+					                    Label="Password" InputType="InputType.Password" Placeholder="password"
+					                    UserAttributes="@(new() { { "autocomplete", "current-password" }, { "aria-required", "true" } })"/>
+				</MudItem>
+				<MudItem md="12">
+					<MudStaticButton Variant="Variant.Filled" Color="Color.Primary" FullWidth="true"
+					                 FormAction="FormAction.Submit">Log in
+					</MudStaticButton>
+				</MudItem>
+			</MudGrid>
+		</EditForm>
+
+	</MudItem>
+</MudGrid>
 
 @code {
 	[Inject] NavigationManager NavigationManager { get; set; }
 	[Inject] IAuthenticationProvider AuthProvider { get; set; }
-	[Inject] IAuthService AuthService { get; set; }
 	[Inject] ISnackbar Snackbar { get; set; }
 	[SupplyParameterFromForm] LoginInputModel Input { get; set; } = new();
 	[SupplyParameterFromQuery] string ReturnUrl { get; set; }
 	[CascadingParameter] HttpContext HttpContext { get; set; } = default!;
 
-	protected override async Task OnAfterRenderAsync(bool firstRender) {
-		Console.WriteLine(HttpContext?.User?.Identity?.IsAuthenticated);
-		if (firstRender && !AuthService.IsLoggedIn) {
-			var restoredFromState = await AuthService.GetStateFromTokenAsync();
-			if (restoredFromState) {
-				Redirect();
-			}
+	// private string errorMessage;
 
+	protected override async Task OnAfterRenderAsync(bool firstRender) {
+		var authenticated = HttpContext.User.Identity?.IsAuthenticated == true;
+		Console.WriteLine(authenticated);
+		if (firstRender && !authenticated) {
 			// Try to log in with a dummy user that will always work in insecure mode
-			var dummy = JwtTokenService.GenerateRandomString(10);
+			var dummy = RandomService.GenerateRandomString(10);
 			var (status, principal) = await LoginUser(dummy, dummy);
 			if (status == HttpAuthenticationRequestStatus.Authenticated) {
-				await AuthService.Login(principal);
+				await HttpContext.SignInAsync(principal);
 				Redirect();
 			}
 		}
@@ -64,13 +73,14 @@
 			Snackbar.Add("Invalid username or password", Severity.Error);
 			return;
 		}
+		await HttpContext.SignInAsync(principal);
 
-		await AuthService.Login(principal);
+		// await AuthService.Login(principal);
 		Redirect();
 	}
 
 	async Task<(HttpAuthenticationRequestStatus, ClaimsPrincipal)> LoginUser(string username, string password) {
-		var request = new HttpAuthenticationRequest(new DefaultHttpContext(), username, password);
+		var request = new HttpAuthenticationRequest(HttpContext, username, password);
 		AuthProvider.Authenticate(request);
 		return await request.AuthenticateAsync();
 	}
