+++ src/KurrentDB/Components/Auth/Logout.razor
@@ -1,12 +1,17 @@
+@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
 @page "/ui/logout"
-@using EventStore.Core.Authentication
+@using Microsoft.AspNetCore.Authentication
+@using Microsoft.AspNetCore.Http
+@attribute [ExcludeFromInteractiveRouting]
 
 @code {
 	[Inject] private NavigationManager NavigationManager { get; set; }
-	[Inject] private IAuthService AuthService { get; set; }
+	[CascadingParameter] HttpContext HttpContext { get; set; } = default!;
 
 	protected override async Task OnAfterRenderAsync(bool firstRender) {
-		await AuthService.Logout();
+		await HttpContext.SignOutAsync();
 		NavigationManager.NavigateTo("/ui/login", true);
 	}
 }
