+++ src/KurrentDB/Components/Cluster/ClusterMember.razor
@@ -1,3 +1,6 @@
+@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
 @using EventStore.Core.Cluster
 @using EventStore.Core.Data
