--- src/KurrentDB/Components/Layout/CheckAuthorization.razor
@@ -1,23 +1,0 @@
-@using EventStore.Core.Authentication
-
-<h1>@_message</h1>
-
-@code {
-    [Inject] private IAuthService AuthService { get; set; }
-    [Inject] private NavigationManager NavigationManager { get; set; }
-    string _returnUrl;
-    string _message = string.Empty;
-
-    protected override void OnInitialized()
-    {
-        _returnUrl = NavigationManager.ToBaseRelativePath(NavigationManager.Uri);
-        if (AuthService.IsLoggedIn)
-        {
-            _message = "Sorry, you are not authorized to view this page.";
-        }
-        else
-        {
-            NavigationManager.NavigateTo($"/ui/login?returnUrl={_returnUrl}", true);
-        }
-    }
-}
