+++ src/KurrentDB/Components/Licensed/LicensedPageCore.razor
@@ -1,14 +1,18 @@
+@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
 @inherits WithLicense
+@attribute [ExcludeFromInteractiveRouting]
+@rendermode InteractiveServer
 
 <PageTitle>KurrentDB: @PageName</PageTitle>
+<MudPopoverProvider/>
 
 @if (HasLicense) {
 	@Body
 } else {
 	<MudText>
-		@PageName page requires a valid
-		<MudLink Href="/ui/license">license</MudLink>
-		.
+		@PageName page requires a valid <MudLink Href="/ui/license">license</MudLink>.
 	</MudText>
 }
