+++ src/KurrentDB/Components/Pages/Config.razor
@@ -1,3 +1,6 @@
+@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
 @page "/ui/config"
 @attribute [Authorize]
 @using EventStore.Core
