+++ src/KurrentDB/Components/Pages/Error.razor
@@ -1,4 +1,7 @@
-﻿@page "/Error"
+﻿@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
+@page "/Error"
 @using System.Diagnostics
 @using Microsoft.AspNetCore.Http
