+++ src/KurrentDB/Components/Pages/License.razor
@@ -1,12 +1,17 @@
+@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
 @page "/ui/license"
+@inherits KurrentDB.Components.ServerSide.ServerSideBase
+@attribute [Authorize]
 @using EventStore.Licensing
 @using EventStore.Plugins.Licensing
 @using JetBrains.Annotations
 @using Microsoft.AspNetCore.Authorization
-@attribute [Authorize]
 
 <PageTitle>KurrentDB: License</PageTitle>
 
+<MudPopoverProvider/>
 <MudCard>
 	<MudCardHeader>
 		<CardHeaderContent>
