+++ src/KurrentDB/Components/Pages/Logging.razor
@@ -1,3 +1,6 @@
+@* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
+@* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
+
 @page "/ui/logs"
 @attribute [Authorize]
 @using System.ComponentModel
@@ -8,7 +11,6 @@
 @inject LogObserver LogObserver
 @inherits LicensedPage
 @implements IDisposable
-
 
 <MudDataGrid @ref="_dataGrid" Items="@LogObserver.Items" Dense="@true" Filterable="true">
 	<Columns>
