+++ src/KurrentDB/Components/ServerSide/ServerSideBase.razor
@@ -1,13 +1,12 @@
 @* Copyright (c) Kurrent, Inc and/or licensed to Kurrent, Inc under one or more agreements. *@
 @* Kurrent, Inc licenses this file to you under the Kurrent License v1 (see LICENSE.md). *@
 
-@inherits WithLicense
+@attribute [ExcludeFromInteractiveRouting]
+@rendermode InteractiveServer
 
-@if (HasLicense) {
-	@ChildContent
-}
+<MudPopoverProvider/>
+@ChildContent
 
 @code {
-	[Parameter]
-    public RenderFragment ChildContent { get; set; }
+	[Parameter] public RenderFragment ChildContent { get; set; }
 }
